+++ app/assets/javascripts/surveyor/surveys.js.coffee
@@ -21,7 +21,7 @@ $(document).ready ->
   $("[data-toggle='tooltip']").tooltip()
 
   ### Survey Table ###
-  $(document).on 'click', '#new-survey-button', '#new-form-button', ->
+  $(document).on 'click', '#new-survey-button, #new-form-button', ->
     type = $(this).data('type')
     klass = $(this).data('klass')
 
@@ -29,7 +29,7 @@ $(document).ready ->
       title: if type == 'Form' then I18n['surveyor']['forms']['new'] else I18n['surveyor']['systemsurveys']['new']
       text: "#{I18n['surveyor']['surveys']['new']['text']} #{klass}."
       input: 'text'
-      inputPlaceholder: 'Access Code'
+      inputPlaceholder: I18n['surveyor']['surveys']['new']['placeholder']
       showCancelButton: true
       inputValidator: (access_code) ->
         return new Promise((resolve) ->
@@ -58,10 +58,11 @@ $(document).ready ->
       showCancelButton: true
       confirmButtonColor: '#DD6B55'
       confirmButtonText: 'Delete'
-    }).then ->
-      $.ajax
-        type: 'delete'
-        url: "/surveyor/surveys/#{survey_id}.js"
+    }).then (data) ->
+      if data['value']
+        $.ajax
+          type: 'delete'
+          url: "/surveyor/surveys/#{survey_id}.js"
 
   ### Survey Modal ###
   $(document).on 'hide.bs.modal', '#modal_place', ->
